+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 * Add `merchant_id` alias to ConnectedMerchantStatusTransitioned and ConnectedMerchantPayPalStatusChanged Auth webhooks
 * Fix webhook testing sample xml for dispute webhooks to include `amount-won` and `amount-disputed`
 * `Config` creation with `AccessToken` should not require `environment`. An `Error` will be logged if optional `environment` does not match with `AccessToken`'s parsed environment
+* Add processor respone code and processor response text to authorization adjustments subfield in transaction response.
 
 ## 2.10.0
 * Bank Account verifications API