+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## Unreleased
 * Add subscription charged unsuccessfully sample webhook to webhook testing gateway
 * Add `processor_response_code` and `processor_response_text` to authorization adjustments subfield in transaction response.
+* Add `Dispute` to top level `braintree` object
 
 ## 2.11.0
 * Throw error if invalid Payload on `CustomerGateway.create`