+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## Unreleased
+## 2.12.0
 * Add subscription charged unsuccessfully sample webhook to webhook testing gateway
 * Add `processor_response_code` and `processor_response_text` to authorization adjustments subfield in transaction response.
 * Add `Dispute` to top level `braintree` object